Moving core from separate jar to codebase, since pom is not able to find system dependency
