+++ pom.xml
@@ -75,12 +75,5 @@
       <artifactId>json-simple</artifactId>
       <version>[1.1.1,)</version>
     </dependency>
-    <dependency>
-      <groupId>core</groupId>
-      <artifactId>aqCore</artifactId>
-      <version>1.0</version>
-      <scope>system</scope>
-      <systemPath>${project.basedir}/src/lib/core-1.0-SNAPSHOT.jar</systemPath>
-    </dependency>
   </dependencies>
 </project>
