+++ src/main/java/com/aq/aqconnect/AQException.java
@@ -1,4 +1,5 @@
-package aqPluginCore;
+package com.aq.aqconnect;
+
 
 public class AQException extends RuntimeException{
