+++ src/main/java/com/aq/aqconnect/AQFormValidate.java
@@ -1,4 +1,5 @@
-package aqPluginCore;
+package com.aq.aqconnect;
+
 
 import java.net.URL;
 import java.util.regex.Pattern;
