+++ src/main/java/com/aq/aqconnect/AQPluginBuilderAction.java
@@ -20,7 +20,6 @@ import javax.annotation.Nonnull;
 import javax.servlet.ServletException;
 import java.io.IOException;
 import java.io.PrintStream;
-import aqPluginCore.*;
 import org.kohsuke.stapler.QueryParameter;
 import org.kohsuke.stapler.verb.POST;
 import org.json.simple.parser.ParseException;
